Add note about missing Purview modules
Some of the modules from the learning path about
Microsoft Purview has disappeard. Don't know if
Purview is still part of the DP-203 syllabus, so
I added a link to the documentation for Purview.
Will look more into this and decide what to do
about this when I get to that part of the
syllabus.